Rev1

+++ app/Http/Controllers/KategoriController.php
@@ -2,63 +2,39 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Barang;
 use App\Models\Kategori;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class KategoriController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    public $lab;
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware(function ($request, $next) {
+            $this->lab = Auth::user()->post;
+            return $next($request);
+        });
+    }
+
     public function index()
     {
-        if (Auth::user()->role_id == 3) {
-            $kategori_lab = 1;
-        } elseif (Auth::user()->role_id == 4) {
-            $kategori_lab = 2;
-        } elseif (Auth::user()->role_id == 5) {
-            $kategori_lab = 3;
-        } elseif (Auth::user()->role_id == 6) {
-            $kategori_lab = 4;
-        }
-        $kategori = Kategori::where('kategori_lab', $kategori_lab)->paginate(5);
+        $kategori = Kategori::where('laboratorium_id', $this->lab)->paginate(5);
         return view('backend.barang.kategori.index', compact('kategori'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         return view('backend.barang.kategori.add');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $request->validate([
             'nama_kategori' => 'required',
         ]);
-
-        if (Auth::user()->role_id == 3) {
-            $kategori_lab = 1;
-        } elseif (Auth::user()->role_id == 4) {
-            $kategori_lab = 2;
-        } elseif (Auth::user()->role_id == 5) {
-            $kategori_lab = 3;
-        } elseif (Auth::user()->role_id == 6) {
-            $kategori_lab = 4;
-        }
 
         $max = Kategori::max('id');
         $kode = $max + 1;
@@ -75,9 +51,10 @@ class KategoriController extends Controller
         }
 
         $kategori = Kategori::create([
-            'nama_kategori' => $request->nama_kategori,
+            'id'  => $kode,
             'kode' => $kode_id,
-            'kategori_lab' => $kategori_lab,
+            'nama_kategori' => $request->nama_kategori,
+            'laboratorium_id' => $this->lab,
         ]);
 
         if ($kategori) {
@@ -87,35 +64,11 @@ class KategoriController extends Controller
         }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Kategori $kategori)
     {
         return view('backend.barang.kategori.edit', compact('kategori'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $id)
     {
         $request->validate([
@@ -132,14 +85,13 @@ class KategoriController extends Controller
         }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
+        $id = $request->delete_id;
+        $barang = Barang::where('kategori_id', $id)->first();
+        if ($barang) {
+            return redirect()->route('kategori.index')->with(['error', 'Kategori Masih digunakan!']);
+        }
         $kategori = Kategori::whereId($id)->delete();
         if ($kategori) {
             return redirect()->route('kategori.index')->with(['success', 'Kategori berhasil dihapus']);
